removed unecessary code
